+++ src/mimic/src/lib.rs
@@ -97,7 +97,6 @@ pub mod prelude {
 ///
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use snafu::Snafu;
 
 ///
 /// ERROR
@@ -105,26 +104,55 @@ use snafu::Snafu;
 /// consolidates all the different crate errors into one place
 ///
 
-#[derive(CandidType, Debug, Serialize, Deserialize, Snafu)]
-pub enum Error {
-    #[snafu(transparent)]
-    Api { source: api::Error },
+pub const ERROR_API: u8 = 101;
+pub const ERROR_CONFIG: u8 = 102;
+pub const ERROR_DB: u8 = 103;
+pub const ERROR_QUERY: u8 = 104;
+pub const ERROR_CORE_SCHEMA: u8 = 105;
+pub const ERROR_CORE_STATE: u8 = 106;
+pub const ERROR_CORE_WASM: u8 = 107;
 
-    #[snafu(transparent)]
-    Config { source: config::Error },
+#[derive(CandidType, Debug, Serialize, Deserialize)]
+pub struct Error(u8, String);
 
-    #[snafu(transparent)]
-    Db { source: db::Error },
+impl From<api::Error> for Error {
+    fn from(error: api::Error) -> Self {
+        Self(ERROR_API, error.to_string())
+    }
+}
 
-    #[snafu(transparent)]
-    Query { source: db_query::Error },
+impl From<config::Error> for Error {
+    fn from(error: config::Error) -> Self {
+        Self(ERROR_CONFIG, error.to_string())
+    }
+}
 
-    #[snafu(transparent)]
-    CoreSchema { source: core_schema::Error },
+impl From<db::Error> for Error {
+    fn from(error: db::Error) -> Self {
+        Self(ERROR_DB, error.to_string())
+    }
+}
 
-    #[snafu(transparent)]
-    CoreState { source: core_state::Error },
+impl From<db::query::Error> for Error {
+    fn from(error: db::query::Error) -> Self {
+        Self(ERROR_QUERY, error.to_string())
+    }
+}
 
-    #[snafu(transparent)]
-    CoreWasm { source: core_wasm::Error },
+impl From<core_schema::Error> for Error {
+    fn from(error: core_schema::Error) -> Self {
+        Self(ERROR_CORE_SCHEMA, error.to_string())
+    }
+}
+
+impl From<core_state::Error> for Error {
+    fn from(error: core_state::Error) -> Self {
+        Self(ERROR_CORE_STATE, error.to_string())
+    }
+}
+
+impl From<core_wasm::Error> for Error {
+    fn from(error: core_wasm::Error) -> Self {
+        Self(ERROR_CORE_WASM, error.to_string())
+    }
 }